backend/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
firebase-admin>=6.0.0,<7.0.0
google-generativeai==0.8.5
python-dotenv==1.1.0
azure-storage-blob>=12.0.0